refactor: reduce code duplication for loss and validation functions
