+++ src/train_refine.py
@@ -2,13 +2,9 @@
 import datetime
 import os
 
-import kornia
 import torch
-from torch import nn
 from torch.cuda.amp import GradScaler, autocast
-from torch.nn import functional as F
 from torch.optim import Adam
-from torch.utils.data import DataLoader
 from torch.utils.tensorboard import SummaryWriter
 from torchvision.utils import make_grid
 from tqdm import tqdm
@@ -16,6 +12,7 @@ from tqdm import tqdm
 from src.model import MattingRefine
 from src.utils.augmentation import random_crop, train_step_augmenter
 from src.utils.generator import define_generators
+from src.utils.train_utils import compute_refine_loss, valid
 
 # check if cuda available
 DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
@@ -90,7 +87,7 @@ class FineMatte:
                         pred_err_sm,
                         _,
                     ) = self.model(true_src, true_bgr)
-                    loss = compute_loss(
+                    loss = compute_refine_loss(
                         pred_pha,
                         pred_fgr,
                         pred_pha_sm,
@@ -128,7 +125,7 @@ class FineMatte:
                 del true_pha, true_fgr, true_src, true_bgr
                 del pred_pha, pred_fgr, pred_pha_sm, pred_fgr_sm, pred_err_sm
 
-            current_val_loss = valid(self.model, valid_loader, writer, step)
+            current_val_loss = valid(self.model, "refine", valid_loader, writer, step)
 
             if current_val_loss < valid_loss:
                 valid_loss = current_val_loss
@@ -136,84 +133,6 @@ class FineMatte:
                     self.model.state_dict(),
                     f"checkpoint/matting_refine/{now}/epoch-{epoch}-loss-{valid_loss:.4f}.pth",
                 )
-
-
-# --------------- Utils ---------------
-
-
-def compute_loss(
-    pred_pha_lg,
-    pred_fgr_lg,
-    pred_pha_sm,
-    pred_fgr_sm,
-    pred_err_sm,
-    true_pha_lg,
-    true_fgr_lg,
-):
-    """loss calculating function"""
-    true_pha_sm = kornia.resize(true_pha_lg, pred_pha_sm.shape[2:])
-    true_fgr_sm = kornia.resize(true_fgr_lg, pred_fgr_sm.shape[2:])
-    true_msk_lg = true_pha_lg != 0
-    true_msk_sm = true_pha_sm != 0
-    return (
-        F.l1_loss(pred_pha_lg, true_pha_lg)
-        + F.l1_loss(pred_pha_sm, true_pha_sm)
-        + F.l1_loss(kornia.sobel(pred_pha_lg), kornia.sobel(true_pha_lg))
-        + F.l1_loss(kornia.sobel(pred_pha_sm), kornia.sobel(true_pha_sm))
-        + F.l1_loss(pred_fgr_lg * true_msk_lg, true_fgr_lg * true_msk_lg)
-        + F.l1_loss(pred_fgr_sm * true_msk_sm, true_fgr_sm * true_msk_sm)
-        + F.mse_loss(
-            kornia.resize(pred_err_sm, true_pha_lg.shape[2:]),
-            kornia.resize(pred_pha_sm, true_pha_lg.shape[2:]).sub(true_pha_lg).abs(),
-        )
-    )
-
-
-def valid(
-    model: nn.Module, dataloader: DataLoader, writer: SummaryWriter, step: int
-) -> float:
-    """model evaluation step executor
-
-    Args:
-        model (nn.Module): [description]
-        dataloader (DataLoader): [description]
-        writer (SummaryWriter): [description]
-        step (int): [description]
-
-    Returns:
-        float: validation loss
-    """
-
-    model.eval()
-    loss_total = 0
-    loss_count = 0
-    with torch.no_grad():
-        for (true_pha, true_fgr), true_bgr in dataloader:
-            batch_size = true_pha.size(0)
-
-            true_pha = true_pha.to(DEVICE)
-            true_fgr = true_fgr.to(DEVICE)
-            true_bgr = true_bgr.to(DEVICE)
-            true_src = true_pha * true_fgr + (1 - true_pha) * true_bgr
-
-            pred_pha, pred_fgr, pred_pha_sm, pred_fgr_sm, pred_err_sm, _ = model(
-                true_src, true_bgr
-            )
-            loss = compute_loss(
-                pred_pha,
-                pred_fgr,
-                pred_pha_sm,
-                pred_fgr_sm,
-                pred_err_sm,
-                true_pha,
-                true_fgr,
-            )
-            loss_total += loss.cpu().item() * batch_size
-            loss_count += batch_size
-
-    writer.add_scalar("valid_loss", loss_total / loss_count, step)
-    model.train()
-    return loss_total / loss_count
 
 
 if __name__ == "__main__":
